+++ test/Hoyo/Test/Utils.hs
@@ -1,12 +1,12 @@
 {-# LANGUAGE TemplateHaskell #-}
 module Hoyo.Test.Utils where
 
-import qualified Data.Text           as T
+import qualified Data.Text       as T
 
 import           Hoyo
-import           Hoyo.Internal.Utils
 import           Hoyo.Test.Gen
 import           Hoyo.Test.Hoyo
+import           Hoyo.Utils
 
 import           Lens.Micro
 
@@ -22,7 +22,7 @@ testMaximumDefaultInt x xs =
 prop_MaximumDefault :: Property
 prop_MaximumDefault = property testMaximumDefaultInt
 
-testReadBool :: T.Text -> Either T.Text Bool -> Property
+testReadBool :: T.Text -> Either HoyoException Bool -> Property
 testReadBool tx expected = readBool tx === expected
 
 prop_ReadBool :: Property
@@ -35,12 +35,12 @@ prop_ReadBool = forAll cases (uncurry testReadBool)
 
     incorrect = do
       tx <- suchThat text (`notElem` (trues <> falses))
-      return (tx, Left ("Couldn't parse bool: " <> tx))
+      return (tx, Left $ ParseException ["Couldn't parse bool: " <> tx])
 
     trues = ["true", "True"]
     falses = ["false", "False"]
 
-testReadInt :: T.Text -> Either T.Text Int -> Property
+testReadInt :: T.Text -> Either HoyoException Int -> Property
 testReadInt tx expected = readInt tx === expected
 
 prop_ReadInt :: Property
@@ -58,26 +58,26 @@ prop_ReadInt = forAll cases (uncurry testReadInt)
       nondig <- elements (['a'..'z'] <> ['A'..'Z'])
       lets2 <- listOf letter
       let tx = T.pack (lets1 <> [nondig] <> lets2)
-      return (tx, Left ("Couldn't parse integer: " <> tx))
+      return (tx, Left $ ParseException ["Couldn't parse integer: " <> tx])
 
-prop_Assert :: T.Text -> Bool -> Property
-prop_Assert errMsg True = testHoyoMonadEqWithEnv
-                              (Bookmarks [])
-                              defaultConfig
-                              (assert errMsg (return True))
-                              (Right ())
+prop_Assert :: HoyoException -> Bool -> Property
+prop_Assert err True = testHoyoMonadEqWithEnv
+                        (Bookmarks [])
+                        defaultConfig
+                        (assert err (return True))
+                        (Right ())
 prop_Assert errMsg False = testHoyoMonadEqWithEnv
                               (Bookmarks [])
                               defaultConfig
                               (assert errMsg (return False))
                               (Left errMsg)
 
-prop_AssertVerbose :: T.Text -> Bool -> Bool -> Property
-prop_AssertVerbose errMsg bool verb = testHoyoMonadEqWithEnv
+prop_AssertVerbose :: HoyoException -> Bool -> Bool -> Property
+prop_AssertVerbose err bool verb = testHoyoMonadEqWithEnv
                                         (Bookmarks [])
                                         (set failOnError verb defaultConfig)
-                                        (assertVerbose errMsg (return bool))
-                                        (if not bool && verb then Left errMsg else Right bool)
+                                        (assertVerbose err (return bool))
+                                        (if not bool && verb then Left err else Right bool)
 
 return []
 utilsTests :: IO Bool